Rename stats pkg to api (#44)
* Add cli flag for purging ts in TSDB

* Use net.dial to ping tsdb

* Check status code of query and return error if it is not 200

---------

+++ pkg/api/updater/tsdb.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
-	"github.com/mahendrapaipuri/ceems/pkg/stats/base"
-	"github.com/mahendrapaipuri/ceems/pkg/stats/models"
+	"github.com/mahendrapaipuri/ceems/pkg/api/base"
+	"github.com/mahendrapaipuri/ceems/pkg/api/models"
 	"github.com/mahendrapaipuri/ceems/pkg/tsdb"
 	"github.com/prometheus/common/model"
 )
@@ -34,6 +34,10 @@ var (
 		"tsdb.data.cutoff.duration",
 		"Compute units (Batch jobs, VMs, Pods) with wall time less than this period will be ignored. By default none will be ignored. Units Supported: y, w, d, h, m, s, ms.",
 	).Default("0s").String()
+	purgeDataTS = base.CEEMSServerApp.Flag(
+		"tsdb.data.purge.ts",
+		"Ignored compute units (Batch jobs, VMs, Pods) will be purged from the TSDB. Admin API must be enabled in TSDB.",
+	).Default("false").Bool()
 	metricLock = sync.RWMutex{}
 	cutoff     = time.Duration(0 * time.Second)
 )
@@ -217,9 +221,11 @@ func (t *tsdbUpdater) Update(startTime time.Time, endTime time.Time, units []mod
 	}
 
 	// Finally delete time series corresponding to ignoredUnits
-	if err := t.deleteTimeSeries(startTime, endTime, ignoredUnits); err != nil {
-		level.Error(t.Logger).
-			Log("msg", "Failed delete ignored units' time series in TSDB", "err", err)
+	if *purgeDataTS {
+		if err := t.deleteTimeSeries(startTime, endTime, ignoredUnits); err != nil {
+			level.Error(t.Logger).
+				Log("msg", "Failed delete ignored units' time series in TSDB", "err", err)
+		}
 	}
 	return units
 }
